Complete User interface and Executor Service
-Create a ExecutorService, when change rocket powers.
-- This approach allows to change rocket powers, (implicit multiple
threads running) and wait until all tasks are finished

-First version of App as UserOption = 4.
--In this option both rockets changes speed at once. I want to keep it
as an example of different approach on Runnables, where you start all
runnables and don´t wait to finish. *******NOTE****** I had to hardcode
the waiting procces with a boolean to fit this in the UserInterface
approach. For real example check commit with the first version of this
app.